// src/app/common/model/local/asegurado-acceder.ts
export class AseguradoAcceder {
  idee: String;
  nss: String;
  agregadoMedico: String;
  curp: String;
  nombre: String;
  apPaterno: String;
  apMaterno: String;
  fechaNacimiento: String;
  edad: String;
  cveUnidadMedicaAdscripcion: String;
  descUnidadMedicaAdscripcion: String;
  cveDelegacionAdscripcion: String;
  descDelegacionAdscripcion: String;
  sexo: String;
  derechoServicioMedico: String;
  vigente: String;
  vigenciaHasta: String;
}