// src/app/common/model/local/notas-medicas.ts
export class NotasMedicas {
  idee: String;
  fechaNotaMedica: String;
  cveCie10: String;
  descDiagnostico: String;
  complementoDiagnostico: String;
  cveCie9: String;
  descProcedimiento: String;
  cveUnidadMedicaAtencion: String;
  descUnidadMedicaAtencion: String;
  cveIdeeFechaAtencion:any;
}